+++ src/interface/IArea.ts
@@ -3,6 +3,5 @@ import { SplitAreaDirective } from "../components/splitArea.directive";
 export interface IArea {
     comp: SplitAreaDirective;
     size: number;
-    pxToSubtract: number;
     order: number;
 }